Generate value-*-stub.h correctly
These files should be automatically generated by `dbus-binding-tool`
when `make` is run, since not doing so can cause bugs when the dbus
interface changes and these files do not.

Such a bug actually occurred when the recent "volume while muted" patch
was merged, which prompted the need for this commit. The old dbus
interface was still in use, even though a new boolean parameter had been
added to it, and so the whole program failed to compile out of the box.